feat(cli): add --schema flag with build-time JSON Schema generation
Adds `orchard --schema` that prints the JSON Schema for `--json` output.
Agents read the schema to know exact wire-format field names (post-camelCase),
enum string values, and nullability before writing jq filters.

Architecture:
- Json* struct definitions live in json_output_types.rs (single source of truth)
- json_output.rs `include!`s them for the runtime binary
- build.rs `include!`s them to generate schema.json at compile time
- --schema embeds schema.json via include_str!; zero runtime cost
- CI check regenerates schema and diffs against committed version to prevent drift

Part of #317.
