+++ .github/workflows/ci.yml
@@ -51,3 +51,19 @@ jobs:
         with:
           components: rustfmt
       - run: cargo fmt --check
+
+  schema:
+    name: Schema Up-To-Date
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@34e114876b0b11c390a56381ad16ebd13914f8d5 # v4
+      - uses: dtolnay/rust-toolchain@29eef336d9b2848a0b548edc03f92a220660cdb8 # stable
+      - uses: Swatinem/rust-cache@e18b497796c12c097a38f9edb9d0641fb99eee32 # v2
+      - name: Regenerate schema
+        run: cargo build -p orchard
+      - name: Verify schema.json matches committed version
+        run: |
+          if ! git diff --exit-code -- crates/orchard/schema.json; then
+            echo "::error::schema.json is out of sync. Run 'cargo build -p orchard' and commit the updated crates/orchard/schema.json."
+            exit 1
+          fi
