+++ crates/orchard/Cargo.toml
@@ -43,6 +43,12 @@ http-body-util = "0.1"
 hmac = "0.12"
 sha2 = "0.10"
 hex = "0.4"
+schemars = "0.8"
+
+[build-dependencies]
+schemars = "0.8"
+serde = { version = "1", features = ["derive"] }
+serde_json = "1"
 
 [dev-dependencies]
 tempfile = "3"
